+++ pyspelling/__main__.py
@@ -20,8 +20,11 @@ def main():
         '--jobs', '-j',
         action='store',
         type=int,
-        default=0,
-        help="Specify the number of spell checker processes to run in parallel."
+        default=None,
+        help=(
+            "Specify the number of spell checker processes to run in parallel. "
+            "Using 0 will utilize the maximum number of cores."
+        )
     )
     parser.add_argument('--config', '-c', action='store', default='', help="Spelling config.")
     parser.add_argument(
@@ -64,7 +67,9 @@ def run(config, **kwargs):
     verbose = kwargs.get('verbose', 0)
     sources = kwargs.get('sources', [])
     debug = kwargs.get('debug', False)
-    jobs = kwargs.get('jobs', 0)
+    jobs = kwargs.get('jobs', None)
+    if jobs is not None and jobs < 0:
+        jobs = 1
     skip_dict_compile = kwargs.get('skip_dict_compile', False)
 
     fail = False